Added missing doc and minor fixes
