+++ src/engine/source/conf/include/conf/cliconf.hpp
@@ -9,102 +9,89 @@
 
 #include <utils/stringUtils.hpp>
 
+/**
+ * @brief Namespace for the configuration manager.
+ *
+ */
 namespace conf
 {
+/**
+ * @brief Class for managing the configuration of the application.
+ *
+ */
 class CliConf
 {
 private:
     CLI::App_p m_app;
 
-    const CLI::Option* getOption(const std::string& key) const
-    {
-        // Modules are nested by dots
-        // Find the module holding the option
-        const CLI::App* module = m_app.get();
+    /**
+     * @brief Get the internal Option object from a key.
+     *
+     * modules are nested by dots.
+     *
+     * @param key The key of the option.
+     * @return const CLI::Option* The internal option.
+     */
+    const CLI::Option* getOption(const std::string& key) const;
 
-        auto splitted = utils::string::split(key, '.');
-        for (auto i = 0; i < splitted.size() - 1; ++i)
-        {
-            module = module->get_subcommand(splitted[i]);
-        }
-
-        std::string cliKey = "--";
-        cliKey += splitted.back();
-
-        return module->get_option(cliKey);
-    }
-
-    CLI::Option* getOption(const std::string& key)
-    {
-        return const_cast<CLI::Option*>(
-            static_cast<const CliConf*>(this)->getOption(key));
-    }
+    /**
+     * @brief Get the internal Option object from a key.
+     *
+     * modules are nested by dots.
+     *
+     * @param key The key of the option.
+     * @return const CLI::Option* The internal option.
+     */
+    CLI::Option* getOption(const std::string& key);
 
 public:
-    explicit CliConf(CLI::App_p app)
-        : m_app {app}
-    {
-    }
+    /**
+     * @brief Construct a new Cli Conf object from a CLI::App_p.
+     *
+     * @param app The CLI::App_p to use.
+     */
+    explicit CliConf(CLI::App_p app);
 
+    /**
+     * @brief Get the value of an option.
+     *
+     * @tparam T The type of the option.
+     * @param key The key of the option.
+     * @return T The value of the option.
+     *
+     * @throw std::runtime_error If the option does not exist or cannot be casted to T.
+     */
     template<typename T>
     T get(const std::string& key) const
     {
         return getOption(key)->as<T>();
     }
 
-    void saveConfiguration(const std::string& path = "") const
-    {
-        std::string pathStr = path.empty() ? get<std::string>("config") : path;
+    /**
+     * @brief Save the current options to a configuration file.
+     *
+     * @param path The path to save the configuration to.
+     *
+     * @throw std::runtime_error If the configuration file cannot be opened.
+     */
+    void saveConfiguration(const std::string& path = "") const;
 
-        auto savePath = std::filesystem::path(pathStr);
+    /**
+     * @brief Get the configuration as a string.
+     *
+     * @return std::string The configuration.
+     */
+    std::string getConfiguration() const;
 
-        std::ofstream ofs;
-        ofs.open(savePath, std::ios::out | std::ios::trunc);
-        ofs << getConfiguration() << std::endl;
-        ofs.close();
-    }
-
-    std::string getConfiguration() const
-    {
-        auto fmtr = m_app->get_subcommand("server")->get_config_formatter();
-        return fmtr->to_config(m_app->get_subcommand("server"), true, true, "server.");
-    }
-
-    void put(const std::string& key, const std::string& value)
-    {
-        auto opt = getOption(key);
-        auto prev = opt->as<std::string>();
-        // Config option does not behave like the rest
-        // Only options with capture variable return the callback
-        if (opt->get_name() == "--config")
-        {
-            throw std::runtime_error(
-                "Cannot modify config file path, restart the application specifying a "
-                "new path with --config 'path'");
-        }
-
-        auto callback = opt->get_callback();
-        std::vector<std::string> args {value};
-        if (callback(args))
-        {
-            try
-            {
-                opt->clear();
-                opt->add_result(value)->as<std::string>();
-            }
-            catch (...)
-            {
-                opt->clear();
-                opt->add_result(prev);
-                throw;
-            }
-        }
-        else
-        {
-            throw std::runtime_error("Invalid value '" + value + "'for option '" + key
-                                     + "'");
-        }
-    }
+    /**
+     * @brief Put a value in the configuration.
+     *
+     * @param key The key of the option.
+     * @param value The value to put in the option.
+     *
+     * @throw std::runtime_error If the option does not exist or cannot be assigned to.
+     */
+    void put(const std::string& key, const std::string& value);
 };
 } // namespace conf
